Split rest into multiple files
